Renames app to legacy_shim for clarity
